publish new tracks for each test case
